introducing fastobjects (decoration), adding jet reclustering code from Giovanni/Claudia, adding METAnalyzer from susyCMG, various cleanup and fixes
