+++ PhysicsTools/Heppy/python/physicsobjects/slow/Photon.py
@@ -1,0 +1,25 @@
+from PhysicsTools.Heppy.physicsobjects.PhysicsObject import *
+
+class Photon(PhysicsObject ):
+
+    '''                                                                                                                                                                                                                                                                return object from the photon 
+    '''
+    def hOVERe(self):
+        return self.physObj.full5x5_hadTowOverEm()
+
+    def r9(self):
+        return self.physObj.full5x5_r9()
+
+    def sigmaIetaIeta(self):
+        return self.physObj.full5x5_sigmaIetaIeta()
+
+    def chargedHadronIso(self):
+        return self.physObj.chargedHadronIso()
+
+    def neutralHadronIso(self):
+        return self.physObj.neutralHadronIso()
+
+    def photonIso(self):
+        return self.physObj.photonIso()
+
+    pass
