+++ PhysicsTools/Heppy/src/classes.h
@@ -29,6 +29,7 @@ namespace {
     heppy::Davismt2 mt2;
     heppy::mt2w_bisect::mt2w mt2wlept;
     heppy::AlphaT alphaT;
+    heppy::ReclusterJets reclusterJets(vector<float> px, vector<float> py, vector<float> pz, vector<float> E, double ktpower, double rparam);
     //  heppy::SimpleElectron fuffaElectron;
     //  ElectronEnergyCalibrator fuffaElectronCalibrator;
     //  heppy::ElectronEPcombinator fuffaElectronCombinator;
