+++ PhysicsTools/Heppy/src/classes_def.xml
@@ -12,6 +12,8 @@
   <class name="heppy::CMGMuonCleanerBySegmentsAlgo" transient="true" />
   <class name="heppy::EGammaMvaEleEstimatorFWLite" transient="true" />
   <class name="heppy::AlphaT" transient="true" />
+  <class name="heppy::ReclusterJets"  transient="true"/>
+
   <!--<class name="heppy::SimpleElectron" transient="true" />-->
   <!--<class name="ElectronEnergyCalibrator" transient="true" />-->
   <!--<class name="heppy::ElectronEPcombinator" transient="true" />-->
